refactor: change breakPoints file name to breakPoints1 to detect file name changes
affects: @medly-components/theme, @medly-components/utils
