+++ packages/theme/src/core/types.ts
@@ -1,6 +1,6 @@
 export * from './avatar/types';
 export * from './breadCrumbs/types';
-export * from './breakpoints/types';
+export * from './breakpoints1/types';
 export * from './button/types';
 export * from './calendar/types';
 export * from './card/types';
